refactor(webkit)!: data stories standardized + table/pick-list emit (event, item)
Part of the #721 split (data): standardized Storybook stories for data components
+ table 'row-click', table-applied-filters 'remove'/'edit', pick-list 'item-double-click'
now emit the DOM event first.

BREAKING CHANGE: these events now emit (event, …) instead of the bare payload.
