+++ .specs/table.md
@@ -90,7 +90,7 @@ const columns = [
     max-height="480px"
     v-model:sorting="sorting"
     v-model:row-selection="rowSelection"
-    @row-click="(row) => console.log(row)"
+    @row-click="(event, row) => console.log(row)"
   >
     <template #title>Workloads</template>
     <template #toolbar>
@@ -190,7 +190,7 @@ Plug a custom filter builder (e.g. AQL) through `Table.Filter`'s `default` slot;
 - `table-toolbar/table-toolbar.vue` — flex container (end-aligned) for toolbar actions placed in the `toolbar` slot; injects the table context for its testId.
 - `table-search/table-search.vue` — context-aware search input. Reads and drives the TanStack `globalFilter` through the injected table context, so it needs no wiring. Prop `placeholder?: string` (default `'Search'`).
 - `table-filter/table-filter.vue` — context-aware advanced-filter control. An `IconButton` funnel trigger opens an anchored **popover** — a floating panel positioned by the `use-placement` composable, Teleported to `body`, dismissed on outside-click or `Escape` (no Drawer, no invented overlay part). The built-in builder is rendered from `fields: FilterField[]` (default `[]`) as a field + operator + value composer (field and operator via `Select`, value via a tokenized `InputText`; typed value editors come through the escape `default` slot) with an Apply button; committing writes `AppliedFilter[]` through the injected context (and the root `v-model:filters`). Applied conditions are surfaced as `Table.AppliedFilters` chips, never rendered inside the popover. Prop `title?: string` (default `'Filters'`). Emits `apply: [filters: AppliedFilter[]]`, `clear: []`. Slots: `default` (scope `{ fields, applied, apply, clear, remove }`) is the **escape hatch** — provide a custom builder (e.g. the console's AQL) and the built-in body is not rendered; `trigger` (scope `{ open, count }`) customizes the trigger visual. Motion: `animate-popup-scale-in` / `animate-popup-scale-out` paired with `motion-reduce:animate-none`.
-- `table-applied-filters/table-applied-filters.vue` — context-aware chip row for the committed structured filters. Renders one webkit `Chip` (`removable`) per `AppliedFilter` (label = `label`, else a formatted `field operator value`), read from the injected context. Emits `remove: [filter: AppliedFilter]` and `edit: [filter: AppliedFilter]` (re-opens the builder pre-filled). Slot `chip` (scope `{ filter, remove }`) overrides one chip. Placed in the `filters` slot.
+- `table-applied-filters/table-applied-filters.vue` — context-aware chip row for the committed structured filters. Renders one webkit `Chip` (`removable`) per `AppliedFilter` (label = `label`, else a formatted `field operator value`), read from the injected context. Emits `remove: [event: MouseEvent, filter: AppliedFilter]` and `edit: [event: MouseEvent | KeyboardEvent, filter: AppliedFilter]` (re-opens the builder pre-filled). Slot `chip` (scope `{ filter, remove }`, where `remove(event)` takes the activation event) overrides one chip. Placed in the `filters` slot.
 - `table-column-selector/table-column-selector.vue` — context-aware column-visibility picker (the UI over the already-wired `columnVisibility`). An `IconButton` trigger opens an anchored **popover** (the same `use-placement` pattern as `Table.Filter`) listing each hideable leaf column with a `Checkbox`, toggling visibility through the injected `table`. Respects `column.getCanHide()` / the column's `hideable`. Prop `title?: string` (default `'Columns'`). Slot `trigger` (scope `{ open }`). Motion: `animate-popup-scale-in` / `animate-popup-scale-out` paired with `motion-reduce:animate-none`.
 - `table-export/table-export.vue` — context-aware export control (an `IconButton`, download glyph). Calls the injected `exportCsv(options?)`; the root emits `export` with the resolved rows (a consumer may intercept for server-side export, otherwise the built-in CSV util downloads the visible, ordered columns and filtered rows). Props `filename?: string` (default `''` → falls back to the root `exportFilename`), `scope?: 'page' | 'filtered' | 'all'` (default `'filtered'`). Slot `trigger` (scope `{ export }`).
 - `table-refresh-button/table-refresh-button.vue` — context-aware refresh control (an `IconButton`, refresh glyph). Calls the injected `reload()`; the root emits `refresh`. Binds `:disabled` to `disabled || loading` (the injected loading flag), so it can't fire mid-fetch. Prop `disabled?: boolean` (default `false`). No bespoke animation — the loading feedback is the skeleton body.
@@ -265,7 +265,7 @@ Plug a custom filter builder (e.g. AQL) through `Table.Filter`'s `default` slot;
 | `update:columnVisibility` | `value: VisibilityState` | v-model:columnVisibility. |
 | `update:globalFilter` | `value: string` | v-model:globalFilter. |
 | `update:pagination` | `value: PaginationState` | v-model:pagination. |
-| `row-click` | `row: Record<string, unknown>` | Fires when a rendered row is clicked (data-driven mode). |
+| `row-click` | `(event: MouseEvent, row: Record<string, unknown>)` | Fires when a rendered row is clicked (data-driven mode). `event` is the click; `row` is the clicked record. |
 | `update:filters` | `value: AppliedFilter[]` | v-model:filters. |
 | `update:state` | `value: TableStateSnapshot` | v-model:state; debounced so bursts of changes coalesce. |
 | `filter-apply` | `filters: AppliedFilter[]` | The filter builder was committed (Apply) — the explicit re-query signal for `manualFiltering`. |
